Cason's suggestions
Talked with Cason on how to get the nav bar to fit in with how the rest of the app is structured.

We made a displayNavBar function that will be added to every screen to show the nav bar.
