+++ BusApp/app/src/main/java/edu/uga/acm/osp/composables/NavBar.kt
@@ -1,22 +1,35 @@
 package edu.uga.acm.osp.composables
 
+import android.annotation.SuppressLint
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.height
 import androidx.compose.material.Badge
 import androidx.compose.material.BadgedBox
 import androidx.compose.material.BottomNavigation
 import androidx.compose.material.BottomNavigationItem
 import androidx.compose.material.Icon
+import androidx.compose.material.Scaffold
 import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.DateRange
+import androidx.compose.material.icons.filled.Home
+import androidx.compose.material.icons.filled.Notifications
+import androidx.compose.material.icons.filled.Search
+import androidx.compose.material.icons.filled.Settings
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment.Companion.CenterHorizontally
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import androidx.navigation.compose.currentBackStackEntryAsState
+import androidx.navigation.compose.rememberNavController
+import edu.uga.acm.osp.nav.Navigation
 
 /*
     Based off of Phillip Lackner's tutorial video.
@@ -65,6 +78,51 @@ fun NavBar(items: List<NavBarItem>, navController: NavController, modifier: Modi
         }
     }
 }
+
+@SuppressLint("UnusedMaterialScaffoldPaddingParameter")
+@Composable
+fun displayNavBar() {
+    val navController = rememberNavController()
+    Scaffold (
+        bottomBar = {
+            NavBar(items = listOf(
+                NavBarItem(
+                    name = "Home",
+                    route = "home_screen",
+                    icon = Icons.Default.Home
+                ),
+                NavBarItem(
+                    name = "Search",
+                    route = "search_screen",
+                    icon = Icons.Default.Search
+                ),
+                NavBarItem(
+                    name = "Alert",
+                    route = "alert_screen",
+                    icon = Icons.Default.Notifications
+                ),
+                NavBarItem(
+                    name = "Planner",
+                    route = "planner_screen",
+                    icon = Icons.Default.DateRange
+                ),
+                NavBarItem(
+                    name = "Settings",
+                    route = "setting_screen",
+                    icon = Icons.Default.Settings
+                ),
+            ),
+                navController = navController,
+                onItemClick = {
+                    navController.navigate(it.route)
+                }
+            )
+        }
+    ) {
+    Text(text = "Testing")
+    }
+}
+
 
 /*
 @Preview(showBackground = true)
