+++ BusApp/app/src/main/java/edu/uga/acm/osp/nav/HomeScreen.kt
@@ -1,22 +1,35 @@
 package edu.uga.acm.osp.nav
 
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
+import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
+import edu.uga.acm.osp.composables.NavBar
 import edu.uga.acm.osp.composables.NotificationComposable
 import edu.uga.acm.osp.composables.TestComposable
+import edu.uga.acm.osp.composables.displayNavBar
 import edu.uga.acm.osp.composables.myButton
+import edu.uga.acm.osp.ui.theme.BulldogRed
 
 @Composable
 fun HomeScreen(navController: NavController) {
     Column(modifier = Modifier.fillMaxSize()) {
+        Text(text = "Top of Page")
+        Column(modifier = Modifier
+            .height(100.dp)
+            .background(color = BulldogRed)
+        ) {
 
+        }
+        displayNavBar()
     }
 }
